Seg fault test for free function
- points were given if the free function gives segmentation fault, but this can be fixed using this

+++ content/assignments/mini-libc/tests/test_multiple_malloc_free.sh
@@ -41,6 +41,11 @@ test_multiple_malloc_free()
     PID=$!
     mem1=$(ps -o vsz --noheader --pid "$PID")
     sleep 2
+
+    if ! ps -p $PID > /dev/null; then
+        exit 1
+    fi
+
     mem2=$(ps -o vsz --noheader --pid "$PID")
     kill "$PID"
     kill -9 "$PID"
